ユーザー情報とツイートの情報を紐づけ

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -6,16 +6,22 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 // 以下を追加
 use App\Models\Tweet;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use App\Services\TweetService;
+use Illuminate\Support\Facades\Auth;   
 
 class DeleteController extends Controller
 {
     /**
      * Handle the incoming request.
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         //
         $tweetId = (int) $request->route('tweetId');
+        if(!$tweetService->checkOwnTweet(Auth::user()->id, $tweetId)){
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
         return redirect()
